+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IBranchService.java
@@ -29,9 +29,9 @@ public interface IBranchService extends IService<Branch>{
     void createBatch(List<Branch> list) ;
     boolean update(Branch et) ;
     void updateBatch(List<Branch> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    Branch get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    Branch get(Long key) ;
     Branch getDraft(Branch et) ;
     boolean checkKey(Branch et) ;
     boolean save(Branch et) ;
@@ -39,8 +39,8 @@ public interface IBranchService extends IService<Branch>{
     Branch sort(Branch et) ;
     Page<Branch> searchCurProduct(BranchSearchContext context) ;
     Page<Branch> searchDefault(BranchSearchContext context) ;
-    List<Branch> selectByProduct(BigInteger id) ;
-    void removeByProduct(BigInteger id) ;
+    List<Branch> selectByProduct(Long id) ;
+    void removeByProduct(Long id) ;
     /**
      *自定义查询SQL
      * @param sql  select * from table where id =#{et.param}
