+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IDocContentService.java
@@ -29,16 +29,16 @@ public interface IDocContentService extends IService<DocContent>{
     void createBatch(List<DocContent> list) ;
     boolean update(DocContent et) ;
     void updateBatch(List<DocContent> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    DocContent get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    DocContent get(Long key) ;
     DocContent getDraft(DocContent et) ;
     boolean checkKey(DocContent et) ;
     boolean save(DocContent et) ;
     void saveBatch(List<DocContent> list) ;
     Page<DocContent> searchDefault(DocContentSearchContext context) ;
-    List<DocContent> selectByDoc(BigInteger id) ;
-    void removeByDoc(BigInteger id) ;
+    List<DocContent> selectByDoc(Long id) ;
+    void removeByDoc(Long id) ;
     /**
      *自定义查询SQL
      * @param sql  select * from table where id =#{et.param}
