+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IEntryService.java
@@ -29,9 +29,9 @@ public interface IEntryService extends IService<Entry>{
     void createBatch(List<Entry> list) ;
     boolean update(Entry et) ;
     void updateBatch(List<Entry> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    Entry get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    Entry get(Long key) ;
     Entry getDraft(Entry et) ;
     boolean checkKey(Entry et) ;
     boolean save(Entry et) ;
