+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IUserService.java
@@ -29,9 +29,9 @@ public interface IUserService extends IService<User>{
     void createBatch(List<User> list) ;
     boolean update(User et) ;
     void updateBatch(List<User> list) ;
-    boolean remove(BigInteger key) ;
-    void removeBatch(Collection<BigInteger> idList) ;
-    User get(BigInteger key) ;
+    boolean remove(Long key) ;
+    void removeBatch(Collection<Long> idList) ;
+    User get(Long key) ;
     User getDraft(User et) ;
     boolean checkKey(User et) ;
     User getByCommiter(User et) ;
